0.9.1.0 增加DB功能 增加Roll点功能 增加色图Lv0持续成功率增加功能

+++ SheepQQBot3/Extension/BotExtensions.cs
@@ -38,7 +38,7 @@ public static class BotExtensions
         processes.ForEach(each => each.Kill());
     }
 
-    public static string GetSetuSuccessPercent(int setuDoushiLv)
+    public static string GetSetuSuccessPercent(long setuDoushiLv)
     {
         var failedSum = 200 + (int)(150 * Math.Pow(setuDoushiLv, 2))
                             + 200 + (int)(150 * Math.Pow(setuDoushiLv, 2))
